requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
boto3==1.40.25
botocore==1.40.25
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.7
dishka==1.6.0
fastapi==0.116.1
GeoAlchemy2==0.18.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
ijson==3.4.0
jmespath==1.0.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
orjson==3.11.3
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==5.2.1
s3transfer==0.13.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
